requirements.txt
beautifulsoup4
pytz
scipy
numpy
pandas
scikit-learn